StudyAI_Backend/requirements.txt
# Core web framework
fastapi>=0.68.1
uvicorn>=0.15.0
pydantic>=1.8.2

# Firebase
firebase-admin>=6.0.0
python-dotenv>=0.19.0

# ML & Text Processing
transformers>=4.40.0
torch>=2.3.0
accelerate>=1.7.0
sentencepiece>=0.2.0
safetensors>=0.4.2

# PDF & Image Processing
pillow>=10.0.0
pdfminer.six>=20221105
pytesseract>=0.3.10
pdf2image>=1.16.3

# Utilities
requests>=2.32.0
python-multipart>=0.0.9
PyYAML>=6.0.1
